Redesign agents page
